iNAS/requirements_pip.txt
absl-py==0.11.0
astor==0.8.1
blessings @ file:///home/conda/feedstock_root/build_artifacts/blessings_1602247950630/work
cached-property==1.5.2
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
gast==0.2.2
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.33.2
h5py==2.10.0
importlib-metadata==2.0.0
joblib==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
nvidia-ml-py3==7.352.0
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.1
protobuf==3.13.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1603570991184/work
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sklearn==0.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
